Move watchOneTransaction to Watcher class
