+++ src/walletSdk/Watcher/index.ts
@@ -3,6 +3,14 @@ import isEqual from "lodash/isEqual";
 import { Anchor } from "walletSdk/Anchor";
 import { TransactionStatus, WatcherResponse } from "./Types";
 
+interface WatchRegistryAsset {
+  [id: string]: boolean;
+}
+
+interface WatchOneTransactionRegistry {
+  [assetCode: string]: WatchRegistryAsset;
+}
+
 interface WatchAllTransactionsRegistry {
   [assetCode: string]: boolean;
 }
@@ -18,7 +26,13 @@ interface TransactionsRegistry {
 export class Watcher {
   private anchor: Anchor;
 
+  private _oneTransactionWatcher: {
+    [assetCode: string]: {
+      [id: string]: ReturnType<typeof setTimeout>;
+    }
+  };
   private _allTransactionsWatcher?: ReturnType<typeof setTimeout>;
+  private _watchOneTransactionRegistry: WatchOneTransactionRegistry;
   private _watchAllTransactionsRegistry: WatchAllTransactionsRegistry;
   private _transactionsRegistry: TransactionsRegistry;
   private _transactionsIgnoredRegistry: TransactionsRegistry;
@@ -26,6 +40,9 @@ export class Watcher {
   constructor(anchor: Anchor) {
     this.anchor = anchor;
 
+    this._oneTransactionWatcher = {};
+    this._allTransactionsWatcher = undefined;
+    this._watchOneTransactionRegistry = {};
     this._watchAllTransactionsRegistry = {};
     this._transactionsRegistry = {};
     this._transactionsIgnoredRegistry = {};
@@ -204,6 +221,131 @@ export class Watcher {
           this._transactionsRegistry[assetCode] = {};
           this._transactionsIgnoredRegistry[assetCode] = {};
           clearTimeout(this._allTransactionsWatcher);
+        }
+      },
+    };
+  }
+
+  /**
+  * Watch a transaction until it stops pending. Takes three callbacks:
+  * * onMessage - When the transaction comes back as pending_ or incomplete.
+  * * onSuccess - When the transaction comes back as completed / refunded / expired.
+  * * onError - When there's a runtime error, or the transaction comes back as
+  * no_market / too_small / too_large / error.
+  */
+  watchOneTransaction(params: {
+    authToken: string;
+    assetCode: string;
+    id: string;
+    onMessage: (transaction) => void;
+    onSuccess: (transaction) => void;
+    onError: (error: any) => void;
+    timeout?: number;
+    isRetry?: boolean;
+    lang?: string;
+  }): WatcherResponse {
+    const {
+      authToken,
+      assetCode,
+      id,
+      onMessage,
+      onSuccess,
+      onError,
+      timeout = 5000,
+      isRetry = false,
+      lang = this.anchor.language,
+    } = params;
+
+    // make sure to initiate registries for the given asset code
+    // to prevent 'Cannot read properties of undefined' errors
+    if(!this._transactionsRegistry[assetCode]) {
+      this._transactionsRegistry[assetCode] = {};
+    }
+    if(!this._oneTransactionWatcher[assetCode]) {
+      this._oneTransactionWatcher[assetCode] = {};
+    }
+
+    // if it's a first run, drop it in the registry for the given asset code
+    if (!isRetry) {
+      this._watchOneTransactionRegistry[assetCode] = {
+        ...(this._watchOneTransactionRegistry[assetCode] || {}),
+        [id]: true,
+      };
+    }
+
+    // do this all asynchronously (since this func needs to return a cancel fun)
+    this.anchor.getTransactionBy({ authToken, id, lang })
+      .then((transaction) => {
+        // make sure we're still watching
+        if (!this._watchOneTransactionRegistry[assetCode]?.[id]) {
+          return;
+        }
+
+        const registeredTransaction = this._transactionsRegistry[assetCode][
+          transaction.id
+        ];
+
+        // if we've had the transaction before, only report if there is a change
+        if (
+          registeredTransaction &&
+          isEqual(registeredTransaction, transaction)
+        ) {
+          return;
+        }
+
+        this._transactionsRegistry[assetCode][transaction.id] = transaction;
+
+        if (
+          transaction.status.indexOf("pending") === 0 ||
+          transaction.status === TransactionStatus.incomplete
+        ) {
+          if (this._oneTransactionWatcher[assetCode][id]) {
+            clearTimeout(this._oneTransactionWatcher[assetCode][id]);
+          }
+
+          this._oneTransactionWatcher[assetCode][id] = setTimeout(() => {
+            this.watchOneTransaction({
+              ...params,
+              isRetry: true,
+            });
+          }, timeout);
+          onMessage(transaction);
+        } else if (
+          [
+            TransactionStatus.completed, 
+            TransactionStatus.refunded,
+            TransactionStatus.expired
+          ].includes(transaction.status)
+        ) {
+          onSuccess(transaction);
+        } else {
+          onError(transaction);
+        }
+      })
+      .catch((e) => {
+        onError(e);
+      });
+
+    return {
+      refresh: () => {
+        // don't do that if we stopped watching
+        if (!this._watchOneTransactionRegistry[assetCode]?.[id]) {
+          return;
+        }
+
+        if (this._oneTransactionWatcher[assetCode][id]) {
+          clearTimeout(this._oneTransactionWatcher[assetCode][id]);
+        }
+
+        this.watchOneTransaction({
+          ...params,
+          isRetry: true,
+        });
+      },
+      stop: () => {
+        if (this._oneTransactionWatcher[assetCode][id]) {
+          this._watchOneTransactionRegistry[assetCode][id] = false;
+          clearTimeout(this._oneTransactionWatcher[assetCode][id]);
         }
       },
     };
